That was a sneaky one... This fixes #351
Basically, the `DoneRunEvent` didn't have the data of the very last
state for the Run, but a previous one, which was saved in the
database. That's the data that is used when the page is loaded *after*
the run has ended.

So the problem was around for a while. This went unnotified all that
time because only the very last `AssertorResult` (or
`ResourceResponse`) was lost. examity's job made the bug visible
because only one web page is found, with many CSS stylesheets.